refatora: separa funções de erro em errors.py
- Cria errors.py com funções específicas de tratamento de erros
- Simplifica utils.py mantendo apenas utilitários gerais
- Atualiza imports em core.py e llm.py
